+++ R/gg_contour_filled.R
@@ -107,7 +107,8 @@ gg_contour_filled <- function(
     title = NULL,
     subtitle = NULL,
     caption = NULL,
-    titles = snakecase::to_sentence_case
+    titles = snakecase::to_sentence_case,
+    flipped = NULL
 ) {
 
   gg_blanket(
@@ -197,6 +198,7 @@ gg_contour_filled <- function(
     subtitle = subtitle,
     caption = caption,
     titles = titles,
+    flipped = flipped,
     ...
   )
 }