+++ R/gg_errorbar.R
@@ -116,7 +116,8 @@ gg_errorbar <- function(
     title = NULL,
     subtitle = NULL,
     caption = NULL,
-    titles = snakecase::to_sentence_case
+    titles = snakecase::to_sentence_case,
+    flipped = NULL
 ) {
 
   gg_blanket(
@@ -206,6 +207,7 @@ gg_errorbar <- function(
     subtitle = subtitle,
     caption = caption,
     titles = titles,
+    flipped = flipped,
     ...
   )
 }