+++ R/gg_function.R
@@ -105,7 +105,8 @@ gg_function <- function(
     title = NULL,
     subtitle = NULL,
     caption = NULL,
-    titles = snakecase::to_sentence_case
+    titles = snakecase::to_sentence_case,
+    flipped = NULL
 ) {
 
   gg_blanket(
@@ -195,6 +196,7 @@ gg_function <- function(
     subtitle = subtitle,
     caption = caption,
     titles = titles,
+    flipped = flipped,
     ...
   )
 }