+++ R/gg_jitter.R
@@ -113,7 +113,8 @@ gg_jitter <- function(
     title = NULL,
     subtitle = NULL,
     caption = NULL,
-    titles = snakecase::to_sentence_case
+    titles = snakecase::to_sentence_case,
+    flipped = NULL
 ) {
 
   gg_blanket(
@@ -203,6 +204,7 @@ gg_jitter <- function(
     subtitle = subtitle,
     caption = caption,
     titles = titles,
+    flipped = flipped,
     ...
   )
 }